+++ PIXI/Polygon.js
@@ -582,6 +582,36 @@ function wrapslice(arr, start, end) {
 }
 
 /**
+ * Helper function for viewablePoints.
+ * Test if an angle is maximum compared to some other angle
+ * @param {PIXI.Point} pt
+ * @param {PIXI.Point} origin
+ * @param {PIXI.Point} center
+ * @param {number} angle
+ * @param {number} maxAngle
+ * @returns {boolean}
+ */
+function testMaxCWAngle(pt, origin, center, angle, maxAngle) {
+  if ( angle < maxAngle ) return false;
+  if ( foundry.utils.orient2dFast(origin, center, pt) > 0 ) return false; // CCW
+
+  // If the angles are equal, pick the closest point.
+  const dist2Between = PIXI.Point.distanceSquaredBetween;
+  if ( angle === maxAngle ) return dist2Between(origin, pt) < dist2Between(origin, pt);
+  return true;
+}
+
+function testMaxCCWAngle(pt, origin, center, angle, maxAngle) {
+  if ( angle < maxAngle ) return false;
+  if ( foundry.utils.orient2dFast(origin, center, pt) < 0 ) return false; // CW
+
+  // If the angles are equal, pick the closest point.
+  const dist2Between = PIXI.Point.distanceSquaredBetween;
+  if ( angle === maxAngle ) return dist2Between(origin, pt) < dist2Between(origin, pt);
+  return true;
+}
+
+/**
  * Returns the points of the polygon that make up the viewable perimeter
  * as seen from an origin.
  * @param {Point} origin                  Location of the viewer, in 2d.
@@ -591,6 +621,7 @@ function wrapslice(arr, start, end) {
  * @returns {Point[]|number[]}
  */
 function viewablePoints(origin, { returnKeys = false, outermostOnly = false } = {}) {
+
   // Viewable point is a line from origin to the point that does not intersect the polygon
   // the outermost key points are the most ccw and cw of the key points.
   // Get the most clockwise and counterclockwise from the origin point that do not intersect.
@@ -601,65 +632,69 @@ function viewablePoints(origin, { returnKeys = false, outermostOnly = false } = 
 
   const pts = [...this.iteratePoints({ close: false })];
 
-  // Store a map of the points so we can convert key to the exact point (no rounding).
-  const keyMap = new Map();
-  const keySet = new Set(pts.map(pt => {
-    const key = pt.key;
-    keyMap.set(key, pt);
-    return key;
-  }));
-  for ( const edge of this.iterateEdges() ) {
-    const { A, B } = edge;
-    for ( const key of keySet ) {
-      const pt = keyMap.get(key);
-      if ( A.equals(pt) || B.equals(pt) ) continue;
-      if ( foundry.utils.lineSegmentIntersects(origin, pt, A, B) ) keySet.delete(key);
+  // Handle degenerate polygons.
+  // Also, if the polygon contains this origin, use all points of the polygon.
+  // Technically possible for the polygon to be complex and have some points blocked but not relevant
+  // for Foundry use cases.
+  const nPoints = pts.length;
+  if ( nPoints < 3 || this.contains(origin.x, origin.y) ) {
+    // Test if we have a single line segment collinear to the origin; keep the closest point.
+    if ( nPoints === 2 && !foundry.utils.orient2dFast(origin, pts[0], pts[1]) ) {
+      if ( PIXI.Point.distanceSquaredBetween(origin, pts[0]) < PIXI.Point.distanceSquaredBetween(origin, pts[1]) ) pts.pop();
+      else pts.shift();
     }
-  }
-  if ( !keySet.size ) {
-    // console.warn(`No viewablePoints from ${origin.x},${origin.y}.`, this);
-    return [];
-  }
 
-  // Measure how far each point is clockwise or counterclockwise from the origin --> center line.
-  // The farthest points are the outermost key points.
+    if ( returnKeys ) return Array.fromRange(pts.length);
+    return pts;
+  }
+  // Find the points with the largest angle from center --> origin --> pt.
+  // These form the viewing triangle between origin and polygon.
+  // Widest points on either side of the polygon must be viewable.
+  // A point not viewable on one side of polygon center would be blocked by a point with a
+  // larger angle. Thus the point with the largest angle must be viewable.
+  // If two points have the same angle, pick the closer. (Edge is collinear with origin.)
+  // Only pick points after segregating into cw/ccw.
+
   const center = this.center;
   let cwPt;
   let ccwPt;
-  let cwScore = Number.POSITIVE_INFINITY;
-  let ccwScore = Number.NEGATIVE_INFINITY;
-  for ( const key of keySet ) {
-    const pt = keyMap.get(key);
-    const score = foundry.utils.orient2dFast(origin, center, pt);
-    if ( score < cwScore ) {
+  let cwIdx = -1;
+  let ccwIdx = -1;
+  let maxCWAngle = Number.NEGATIVE_INFINITY;
+  let maxCCWAngle = Number.NEGATIVE_INFINITY;
+
+  for ( let i = 0; i < nPoints; i += 1 ) {
+    const pt = pts[i];
+    const angle = PIXI.Point.angleBetween(center, origin, pt);
+    if ( testMaxCWAngle(pt, origin, center, angle, maxCWAngle) ) {
       cwPt = pt;
-      cwScore = score;
+      maxCWAngle = angle;
+      cwIdx = i;
     }
-    if ( score > ccwScore ) {
+    if ( testMaxCCWAngle(pt, origin, center, angle, maxCCWAngle) ) {
       ccwPt = pt;
-      ccwScore = score;
+      maxCCWAngle = angle;
+      ccwIdx = i;
     }
   }
 
-  // Points are from cwPt to ccwPt
-  // Can use `===` b/c we stored the original points in the map.
-  if ( cwPt === ccwPt ) {
-    // Should never happen?
-    // console.warn(`Only one viewablePoint from ${origin.x},${origin.y}.`, this);
-    return [cwPt];
-  }
-  const ccwIdx = pts.indexOf(ccwPt);
-  const cwIdx = pts.indexOf(cwPt);
-  if ( !(~ccwIdx && ~cwIdx) ) {
-    // Should never happen.
-    // console.warn(`No viewablePoints from ${origin.x},${origin.y}.`, this);
+  // Given the starting max angles, should not happen.
+  if ( !cwPt ) cwPt = ccwPt;
+  if ( !ccwPt ) ccwPt = cwPt;
+
+  // Should have defined both by now.
+  if ( !cwPt ) {
+    console.warn("viewablePoints|Points not found", this);
     return [];
   }
 
+  if ( cwPt.equals(ccwPt) ) return returnKeys ? [cwIdx] : [cwPt];
+
   if ( outermostOnly ) {
     if ( returnKeys ) return [cwIdx, ccwIdx];
-    return [pts[cwIdx], pts[ccwIdx]];
+    return [cwPt, ccwPt];
   }
+
   if ( returnKeys ) {
     // Sequentially number the indices, from cwIdx to ccwIdx. Make sure to wrap counting if needed.
     const ln = pts.length;