+++ Changelog.md
@@ -1,3 +1,7 @@
+## 0.2.16
+Avoid WeilerAtherton in constrained token border for now as it returns failed polygons when a wall intersects at a border point.
+Refactor `PIXI.Polygon.prototype.viewablePoints` to avoid intersection testing by taking the points with the largest angles as CW and CCW. This is faster and hopefully more accurate. Test for points contained within the polygon and for degenerate polygons.
+
 ## 0.2.15
 Add checks to ensure that when measuring distance between points, missing x, y, or z properties will be treated as 0.
 Switch to faster `||` when testing for missing z properties. Also catches NaN values.